+++ src/app/components/lineage-graph/lineage-graph.component.ts
@@ -1,28 +1,32 @@
 import {
     ChangeDetectionStrategy,
     Component,
+    ElementRef,
     EventEmitter,
+    HostListener,
     Input,
     OnChanges,
     OnInit,
     Output,
     SimpleChange,
     SimpleChanges,
+    ViewChild,
 } from "@angular/core";
 import { Node, Edge, MiniMapPosition } from "@swimlane/ngx-graph";
-import { DatasetKind } from "src/app/api/kamu.graphql.interface";
+import { DatasetKind, DatasetLineageBasicsFragment } from "src/app/api/kamu.graphql.interface";
 import AppValues from "src/app/common/app.values";
 import { LineageGraphNodeKind } from "src/app/dataset-view/additional-components/lineage-component/lineage-model";
 
 @Component({
     selector: "app-lineage-graph",
     templateUrl: "./lineage-graph.component.html",
+    styleUrls: ["./lineage-graph.component.scss"],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class LineageGraphComponent implements OnChanges, OnInit {
-    @Input() public view: [number, number];
     @Input() public links: Edge[];
     @Input() public nodes: Node[];
+    @Input() public currentDataset: DatasetLineageBasicsFragment;
 
     @Output() public onClickNodeEvent = new EventEmitter<Node>();
 
@@ -42,9 +46,29 @@ export class LineageGraphComponent implements OnChanges, OnInit {
     public readonly DEFAULT_AVATAR_URL = AppValues.DEFAULT_AVATAR_URL;
     public readonly LineageGraphNodeKind: typeof LineageGraphNodeKind = LineageGraphNodeKind;
     public readonly DatasetKind: typeof DatasetKind = DatasetKind;
+    public INITIAL_GRAPH_VIEW_HEIGHT: number = screen.height - 400;
+    public INITIAL_GRAPH_VIEW_WIDTH: number = window.innerWidth - 120;
+    public view: [number, number] = [this.INITIAL_GRAPH_VIEW_WIDTH, this.INITIAL_GRAPH_VIEW_HEIGHT];
+    public showSidePanel = false;
+
+    @ViewChild("containerRef", { static: true }) element: ElementRef;
+    @HostListener("window:resize")
+    private checkWindowSize(): void {
+        this.changeLineageGraphView();
+    }
 
     public ngOnInit(): void {
         this.graphNodes = this.nodes;
+    }
+
+    public changeLineageGraphView(): void {
+        const containerGraph = this.element.nativeElement as HTMLDivElement;
+        const styleElement: CSSStyleDeclaration = getComputedStyle(containerGraph);
+        this.view[0] =
+            containerGraph.clientWidth -
+            parseInt(styleElement.paddingLeft, 10) -
+            parseInt(styleElement.paddingRight, 10);
+        this.view[1] = this.INITIAL_GRAPH_VIEW_HEIGHT;
     }
 
     public ngOnChanges(changes: SimpleChanges): void {
@@ -56,5 +80,9 @@ export class LineageGraphComponent implements OnChanges, OnInit {
 
     public onClickNode(node: Node): void {
         this.onClickNodeEvent.emit(node);
+    }
+
+    public onClickInfo(): void {
+        this.showSidePanel = !this.showSidePanel;
     }
 }