+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Added side panel for lineage graph
+
 ## [0.10.1] - 2023-09-20
 ### Fixed
 - kamu-web-ui#158: pagination broken on profile datasets page